beam_deflection changed
Read the comment before the BD function for the list of inputs. If it is a circle B and H are still 0 and vice versa for R if it is a square. Supports are either fix or tri (really creative names hehe). C is the B distance in the diagrams. It is named C since we used B already for the cross-section. This can be a bit confusing so just check the placement of those. Any questions, just message me. Joe the E is the elastic modulus.